Rename FilterUtils to SortUtils
The tests were failing because it seems like some of the output from the sort tests is polluting the parse tests. I've tried to play around with invalidating the cache with no luck after a certain amount of time. I want to focus on the submission first and the naming keeps the test run order in a way that cross pollution doesn't matter (i.e. sort tests after parse tests). Hopefully I can look at this later.