Null check against the filterList
This is actually a problem with the test. It's not mounted in the test before being run so it needs to be mocked out. Unfortunately it's taking me a while to figure out how to use Jest to stub out components so I'm going to park it and come back to it.